id type changed at StockDao interface to long

+++ src/main/java/com/orgabor/sportsgearstore/dao/StockDao.java
@@ -7,9 +7,9 @@ import com.orgabor.sportsgearstore.products.Product;
 
 public interface StockDao {
 	public void addProduct(Product product);
-	public Product getProduct(int productId);
+	public Product getProduct(long productId);
 	public List<Product> retrieveByCategory(Categories category);
 	public List<Product> retrieveByExpression(String expression);
-	public void updateProduct(int productId, Product updatedProduct);
-	public void deleteProduct(int productId);
+	public void updateProduct(long productId, Product updatedProduct);
+	public void deleteProduct(long productId);
 }
